Allow specifying alternative VS Code settings dirs
Notably adds support for VS Codium and VS Code insiders.

Resolves #34, #29
