+++ main.go
@@ -11,7 +11,6 @@ import (
 	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 	"time"
@@ -36,6 +35,11 @@ func main() {
 		fmt.Printf(`Usage: [-skipsync] %v HOST [DIR] [SSH ARGS...]
 
 Start code-server over SSH.
+
+Environment variables:
+		`+vsCodeConfigDirEnv+`	use special VS Code settings dir.
+		`+vsCodeExtensionsDirEnv+`	use special VS Code extensions dir.
+
 More info: https://github.com/codercom/sshcode
 `, os.Args[0],
 		)
@@ -300,28 +304,4 @@ func rsync(src string, dest string, sshFlags string, excludePaths ...string) err
 	}
 
 	return nil
-}
-
-func configDir() (string, error) {
-	var path string
-	switch runtime.GOOS {
-	case "linux":
-		path = os.ExpandEnv("$HOME/.config/Code/User/")
-	case "darwin":
-		path = os.ExpandEnv("$HOME/Library/Application Support/Code/User/")
-	default:
-		return "", xerrors.Errorf("unsupported platform: %s", runtime.GOOS)
-	}
-	return filepath.Clean(path), nil
-}
-
-func extensionsDir() (string, error) {
-	var path string
-	switch runtime.GOOS {
-	case "linux", "darwin":
-		path = os.ExpandEnv("$HOME/.vscode/extensions/")
-	default:
-		return "", xerrors.Errorf("unsupported platform: %s", runtime.GOOS)
-	}
-	return filepath.Clean(path), nil
 }
